refactor with typescrtipt
