+++ package.json
@@ -4,7 +4,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node dist/index.js",
+    "build": "tsc",
+    "dev": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/index.ts"
   },
   "author": "",
   "license": "ISC",
@@ -16,6 +19,13 @@
     "randomstring": "^1.2.3"
   },
   "devDependencies": {
-    "sequelize-cli": "^6.6.0"
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.12",
+    "@types/node": "^16.11.7",
+    "@types/pg": "^8.11.5",
+    "@types/randomstring": "^1.3.0",
+    "nodemon": "^2.0.15",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.4"
   }
 }
